backend1/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
drf-yasg==1.21.7
django-simple-history==3.4.0
